final commit

+++ src/main/java/com/boottool/manager_system/services/ProductsRepository.java
@@ -3,5 +3,4 @@ package com.boottool.manager_system.services;
 import com.boottool.manager_system.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
-}
+public interface ProductsRepository extends JpaRepository<Product, Long> { }
